Improved Objects
-Created a 2D vector to determine object vectors.
-Balls now start at a straight vertical vector at launch UNLESS the paddle is moving.in which case the ball wil launch horizontally as well at a percentage of the paddles horizontal speed.
-The ball's x component of the vector is updated based on the paddle vector when the ball and paddle collide allowing the ball and miniballs now to move at different angles.

+++ Source/Objects/PowerUpFactory/PowerUps/ExtraLifePowerUp.cs
@@ -32,7 +32,7 @@ namespace Breakout
       //  N/A
       //
       //*********************************************************************************************************************************************
-      public ExtraLifePowerUp(int theCoordinateX, int theCoordinateY) :
+      public ExtraLifePowerUp(float theCoordinateX, float theCoordinateY) :
       base(Image.FromFile("../../../Images/ExtraLifePowerUp.png"), theCoordinateX, theCoordinateY)
       {
       }
